2.0.0
Fully redid the plugin. The plugin (with CommandBlockerLegacy for versions below 1.8.8) works from 1.4.7 to 1.16.2+ now. Automatically backs up your previous configurations to ensure nothing unexpected happens.

The plugin now includes a GUI in order to easily manage all new and existing commands. There is a log file to make sure nothing shady is done, such as an administrator removing commands which should not be removed. 

Overall, the plugin works better and faster than ever before with the thanks of Maven.

** I have also included bStats charts which show how many commands are blocked from operators and normal players **

Some other additions such as the ability to whitelist a player from being blocked from using a command are added.